Adds ability to use a gaussian prior on logAge
logAge now works as the other variables, with a prior mean and sigma in the YAML file corresponding to parameters for a normally distributed prior. The ability to use the original uniform distribution is maintained, and can be activated by setting sigma:logAge to .inf (this is the default setting).