Adds ability to set an artificial max/min age for a model.
The primary intent of this feature is to provide the ability to set a
maximum/minimum age for a model independent of the actual bounds of the model
data. Among other things, this could be convenient for testing issues that only
appear at the bounds of the model data. It could potentially be used to trigger
interpolation above the model, though I've not tested it.

Everything about this feature should be considered experiemental, which is why
there is only the API. If you want to use this, putting e.g.,
`evoModels.mainSequenceEvol->setArtificialMaxAge(13.5);` after the call to
restrictFilters in single or multiPopMcmc is probably a good bet.