Feature: Add support for parallax distance units
Adds support for starting values and priors specified in parallax. This DOES NOT
change the internal model used by BASE-9 (which still runs entirely in distMod),
but it does change how cluster priors are calculated when parallax is specified.

Note that all priors and starting values must be declared for the same parameter
(distMod or parallax) whether they are specified in the .YAML or from the
command line.

 - Cleanup and add beginning parallax support to densities
 - Tweak formatting
 - Refreshes the --help message formatting
 - Wrap up mod -> parallax tranformation with David S.
 - Implement new setting for parallax
 - Clarify help message